INC-119 refactor: getting last month notifications refactored to function getLastMonthNotifications. Added logic to load pages dynamic
